requirements.txt
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
black==23.1.0
classify-imports==4.2.0
click==8.1.3
decorator==5.1.1
Django==4.1.7
djangorestframework==3.14.0
executing==1.2.0
Faker==17.6.0
flake8==6.0.0
flake8-type-checking==2.3.1
ipython==8.11.0
isort==5.12.0
jedi==0.18.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mixins==0.1.4
mypy-extensions==1.0.0
packaging==23.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.1.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
tomli==2.0.1
traitlets==5.9.0
wcwidth==0.2.6
